requirements.txt
streamlit
streamlit-drawable-canvas
tensorflow
opencv-python
Pillow
numpy
